Added a "dynamic" symmetric matrix class (dimension 2 or 3).

+++ Fields.hh
@@ -351,9 +351,9 @@ public:
         return *this;
     }
 
-	// MHS on Nov 3 3015
-   	SymmetricMatrixField &operator/=(ScalarField<Real> scalars) {
-        assert(scalars.domainSize() == m_values.cols());
+    // MHS on Nov 3 3015
+    SymmetricMatrixField &operator/=(ScalarField<Real> scalars) {
+        assert(scalars.domainSize() == size_t(m_values.cols()));
         for (size_t i = 0; i < scalars.domainSize(); ++i)
             m_values.col(i) /= scalars(i);
         return *this;
